#158[FIX] Refatoring Server Artist Fetching
