+++ Setlist.xcworkspace/xcuserdata/a_mcflurry.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -36,68 +36,5 @@
             landmarkType = "0">
          </BreakpointContent>
       </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "7B875AE1-B593-401B-BFE1-2C8C4DB1D673"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "Projects/Feature/Scenes/SearchScene/ViewModel/SearchViewModel.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "44"
-            endingLineNumber = "44"
-            landmarkName = "fetchData()"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "7B875AE1-B593-401B-BFE1-2C8C4DB1D673 - 3fa2ef5cbc4dd03b"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "closure #1 @Sendable (Swift.Optional&lt;Foundation.Data&gt;, Swift.Optional&lt;__C.NSURLResponse&gt;, Swift.Optional&lt;Swift.Error&gt;) -&gt; () in Feature.SearchViewModel.fetchData() -&gt; ()"
-                  moduleName = "Feature"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/a_mcflurry/Documents/Setlist/158/MacC-Team12-Creative8/Projects/Feature/Scenes/SearchScene/ViewModel/SearchViewModel.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "50"
-                  endingLineNumber = "50"
-                  offsetFromSymbolStart = "524">
-               </Location>
-               <Location
-                  uuid = "7B875AE1-B593-401B-BFE1-2C8C4DB1D673 - 7edab606f3c4ae91"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "Feature.SearchViewModel.fetchData() -&gt; ()"
-                  moduleName = "Feature"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/a_mcflurry/Documents/Setlist/158/MacC-Team12-Creative8/Projects/Feature/Scenes/SearchScene/ViewModel/SearchViewModel.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "48"
-                  endingLineNumber = "48"
-                  offsetFromSymbolStart = "396">
-               </Location>
-               <Location
-                  uuid = "7B875AE1-B593-401B-BFE1-2C8C4DB1D673 - 3fa2ef5cbc4dd3f5"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "closure #1 @Sendable (Swift.Optional&lt;Foundation.Data&gt;, Swift.Optional&lt;__C.NSURLResponse&gt;, Swift.Optional&lt;Swift.Error&gt;) -&gt; () in Feature.SearchViewModel.fetchData() -&gt; ()"
-                  moduleName = "Feature"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/a_mcflurry/Documents/Setlist/158/MacC-Team12-Creative8/Projects/Feature/Scenes/SearchScene/ViewModel/SearchViewModel.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "48"
-                  endingLineNumber = "48"
-                  offsetFromSymbolStart = "340">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
    </Breakpoints>
 </Bucket>
